fix: move dynamic pages into clients try to fix offline
