+++ frontend/package.json
@@ -10,7 +10,7 @@
     "lint": "eslint",
     "format:check": "prettier --check .",
     "format:write": "prettier --write .",
-    "prepare": "if [ \"$VERCEL\" != \"1\" ]; then husky; fi",
+    "prepare": "node -e \"if (process.env.VERCEL !== '1') { require('child_process').execSync('husky', {stdio: 'inherit'}) }\"",
     "test": "vitest",
     "cypress": "cypress open",
     "test:e2e": "start-server-and-test dev http://localhost:3000 cypress"
