+++ frontend/src/sw.ts
@@ -46,6 +46,29 @@ const serwist = new Serwist({
       handler: new NetworkFirst({
         cacheName: 'pages',
         networkTimeoutSeconds: 3,
+        runtimeCaching: [
+          {
+            matcher: ({ url }) =>
+              url.pathname.startsWith('/songs/') || url.pathname.startsWith('/playlists/'),
+            handler: new NetworkFirst({
+              cacheName: 'dynamic-pages',
+              networkTimeoutSeconds: 2,
+              plugins: [
+                new CacheableResponsePlugin({ statuses: [0, 200] }),
+                new ExpirationPlugin({ maxEntries: 200, maxAgeSeconds: 30 * 24 * 60 * 60 }),
+              ],
+            }),
+          },
+          {
+            matcher: ({ request }) => request.mode === 'navigate',
+            handler: new NetworkFirst({
+              cacheName: 'pages',
+              networkTimeoutSeconds: 3,
+              plugins: [new CacheableResponsePlugin({ statuses: [0, 200] })],
+            }),
+          },
+          ...defaultCache,
+        ],
         plugins: [new CacheableResponsePlugin({ statuses: [0, 200] })],
       }),
     },
